Overridden method eat() to return the test result in DomesticCat.java. All tests Passed

+++ src/main/java/com/techreturners/cats/Cat.java
@@ -9,7 +9,7 @@ interface Cat {
 
     String getSetting();
 
-    int getAverageHeight();
+    float getAverageHeight();
 
     String eat();
 }
